requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.3
django-distill==3.2.7
django-taggit==6.1.0
idna==3.10
marko==2.1.2
psycopg2-binary==2.9.10
requests==2.32.3
sqlparse==0.5.1
urllib3==2.3.0
